% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNASeq_datasets.R
\docType{data}
\name{miRNASeq}
\alias{miRNASeq}
\alias{ACC.miRNASeq}
\alias{ACC.miRNASeq,BLCA.miRNASeq,BRCA.miRNASeq,CESC.miRNASeq,CHOL.miRNASeq,COAD.miRNASeq,COADREAD.miRNASeq,DLBC.miRNASeq,ESCA.miRNASeq,FPPP.miRNASeq,GBM.miRNASeq,GBMLGG.miRNASeq,HNSC.miRNASeq,KICH.miRNASeq,KIPAN.miRNASeq,KIRC.miRNASeq,KIRP.miRNASeq,LAML.miRNASeq,LGG.miRNASeq,LIHC.miRNASeq,LUAD.miRNASeq,LUSC.miRNASeq,MESO.miRNASeq,OV.miRNASeq,PAAD.miRNASeq,PCPG.miRNASeq,PRAD.miRNASeq,READ.miRNASeq,SARC.miRNASeq,SKCM.miRNASeq,STAD.miRNASeq,STES.miRNASeq,TGCT.miRNASeq,THCA.miRNASeq,THYM.miRNASeq,UCEC.miRNASeq,UCS.miRNASeq,UVM.miRNASeq}
\alias{BLCA.miRNASeq}
\alias{BRCA.miRNASeq}
\alias{CESC.miRNASeq}
\alias{CHOL.miRNASeq}
\alias{COAD.miRNASeq}
\alias{COADREAD.miRNASeq}
\alias{DLBC.miRNASeq}
\alias{ESCA.miRNASeq}
\alias{FPPP.miRNASeq}
\alias{GBM.miRNASeq}
\alias{GBMLGG.miRNASeq}
\alias{HNSC.miRNASeq}
\alias{KICH.miRNASeq}
\alias{KIPAN.miRNASeq}
\alias{KIRC.miRNASeq}
\alias{KIRP.miRNASeq}
\alias{LAML.miRNASeq}
\alias{LGG.miRNASeq}
\alias{LIHC.miRNASeq}
\alias{LUAD.miRNASeq}
\alias{LUSC.miRNASeq}
\alias{MESO.miRNASeq}
\alias{OV.miRNASeq}
\alias{PAAD.miRNASeq}
\alias{PCPG.miRNASeq}
\alias{PRAD.miRNASeq}
\alias{READ.miRNASeq}
\alias{SARC.miRNASeq}
\alias{SKCM.miRNASeq}
\alias{STAD.miRNASeq}
\alias{STES.miRNASeq}
\alias{TGCT.miRNASeq}
\alias{THCA.miRNASeq}
\alias{THYM.miRNASeq}
\alias{UCEC.miRNASeq}
\alias{UCS.miRNASeq}
\alias{UVM.miRNASeq}
\title{miRNASeq datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.miRNASeq

BLCA.miRNASeq

BRCA.miRNASeq

CESC.miRNASeq

CHOL.miRNASeq

COAD.miRNASeq

COADREAD.miRNASeq

DLBC.miRNASeq

ESCA.miRNASeq

FPPP.miRNASeq

GBM.miRNASeq

GBMLGG.miRNASeq

HNSC.miRNASeq

KICH.miRNASeq

KIPAN.miRNASeq

KIRC.miRNASeq

KIRP.miRNASeq

LAML.miRNASeq

LGG.miRNASeq

LIHC.miRNASeq

LUAD.miRNASeq

LUSC.miRNASeq

MESO.miRNASeq

OV.miRNASeq

PAAD.miRNASeq

PCPG.miRNASeq

PRAD.miRNASeq

READ.miRNASeq

SARC.miRNASeq

SKCM.miRNASeq

STAD.miRNASeq

STES.miRNASeq

TGCT.miRNASeq

THCA.miRNASeq

THYM.miRNASeq

UCEC.miRNASeq

UCS.miRNASeq

UVM.miRNASeq
}
\value{
Data frames with miRNASeq data.
}
\description{
Package provides miRNASeq datasets from The Cancer Genome Atlas Project for
all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for
the date: \code{2015-11-01}. The process is described here:
\url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.data}
family can be found here: \code{browseVignettes("RTCGA")}. miRNASeq data format
is explained here \url{https://wiki.nci.nih.gov/display/TCGA/miRNASeq#miRNASeq-DataOverview}.
There is one extra column "machine" in each dataset which describes a machine that 
produced dataset. It can be: Illumina Genome Analyzer, Illumina HiSeq 2000 or both.
Converting \pkg{RTCGA.miRNASeq} datasets from \code{data.frames} to Bioconductor classes is explained here \link{convertTCGA}.
}
\details{
\code{browseVignettes("RTCGA")}
}
\examples{

\dontrun{
browseVignettes("RTCGA")
}


}
\keyword{datasets}
